#925 Refactoring of nrt scripts to avoid having to define useless environment variables when reusing nrt script features in others scripts
