+++ dynawo/sources/Launcher/main.cpp
@@ -118,14 +118,14 @@ int main(int argc, char ** argv) {
     launchSimu(jobsFileName);
   } catch (const DYN::Error& e) {
     return e.type();
-  } catch (const char *s) {
+  } catch (const char*) {
     return -1;
-  } catch (const string & Msg) {
+  } catch (const string&) {
     return -1;
   } catch (const xml::sax::parser::ParserException& exp) {
     Trace::error() << DYNLog(XmlParsingError, jobsFileName, exp.what()) << Trace::endline;
     return -1;
-  } catch (std::exception & exc) {
+  } catch (std::exception&) {
     return -1;
   } catch (...) {
     Trace::error() << __FILE__ << " " << __LINE__ << " " << DYNLog(UnexpectedError) << Trace::endline;
